move resources.py to /inventory
