+++ google/cloud/security/inventory/resources.py
